GimmickManager 옮김
