--- Assets/1_Append/LHS/GimmickManager.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 8cca48926e002684291cb876a6902f80
